+++ software/src/modules/wifi/wifi.cpp
@@ -330,7 +330,7 @@ bool Wifi::apply_sta_config_and_connect(WifiState current_state)
     dns.fromString(sta_config_in_use.get("dns")->asEphemeralCStr());
     dns2.fromString(sta_config_in_use.get("dns2")->asEphemeralCStr());
 
-    if (ip != 0) {
+    if ((uint32_t)ip != 0) {
         WiFi.config(ip, gateway, subnet, dns, dns2);
     } else {
         WiFi.config((uint32_t)0, (uint32_t)0, (uint32_t)0);
@@ -498,7 +498,7 @@ void Wifi::setup()
 
                 // FIXME: Use a better way of time keeping here.
                 if (connected_for < 0x7FFFFFFF)
-                    logger.printfln("Disconnected from %s: %s (%u). Was connected for %u seconds.", sta_config_in_use.get("ssid")->asEphemeralCStr(), reason, reason_code, connected_for / 1000);
+                    logger.printfln("Disconnected from %s: %s (%u). Was connected for %lu seconds.", sta_config_in_use.get("ssid")->asEphemeralCStr(), reason, reason_code, connected_for / 1000);
                 else
                     logger.printfln("Disconnected from %s: %s (%u). Was connected for a long time.", sta_config_in_use.get("ssid")->asEphemeralCStr(), reason, reason_code);
 
@@ -740,7 +740,7 @@ void Wifi::get_scan_results(StringBuilder *sb, int network_count)
             sb->putc(',');
         }
 
-        sb->printf("{\"ssid\":\"%s\",\"bssid\":\"%s\",\"rssi\":%d,\"channel\":%d,\"encryption\":%d}",
+        sb->printf("{\"ssid\":\"%s\",\"bssid\":\"%s\",\"rssi\":%ld,\"channel\":%ld,\"encryption\":%d}",
                    WiFi.SSID(i).c_str(), WiFi.BSSIDstr(i).c_str(), WiFi.RSSI(i), WiFi.channel(i), WiFi.encryptionType(i));
     }
